requirements.txt
dj-database-url==0.4.0
dj-static==0.0.6
Django==1.10.4
django-crispy-forms==1.6.0
gunicorn==19.4.5
heroku==0.1.4
httplib2==0.9.2
psycopg2==2.6.1
PySocks==1.5.7
pystache==0.5.4
python-dateutil==1.5
pytz==2016.3
requests==2.9.1
six==1.10.0
South==1.0.2
static==1.1.1
static3==0.6.1
twilio==5.4.0
virtualenv==14.0.6
whitenoise==2.0.6
django-multiselectfield
djangorestframework